Migrate quickcheck tests to implementations
