+++ spinoso-string/src/lib.rs
@@ -2358,147 +2358,8 @@ fn chomp(string: &mut String, separator: Option<&[u8]>) -> bool {
 #[cfg(test)]
 #[allow(clippy::invisible_characters)]
 mod tests {
-    use alloc::vec::Vec;
-    use core::str;
-
-    use quickcheck::quickcheck;
-
-    use crate::String;
     use crate::center::CenterError;
-
-    quickcheck! {
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_char_len_utf8_contents_utf8_string(contents: alloc::string::String) -> bool {
-            let expected = contents.chars().count();
-            let s = String::utf8(contents.into_bytes());
-            s.char_len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_len_utf8_contents_utf8_string(contents: alloc::string::String) -> bool {
-            let expected = contents.len();
-            let s = String::utf8(contents.into_bytes());
-            s.len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_bytesize_utf8_contents_utf8_string(contents: alloc::string::String) -> bool {
-            let expected = contents.len();
-            let s = String::utf8(contents.into_bytes());
-            s.bytesize() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_char_len_binary_contents_utf8_string(contents: Vec<u8>) -> bool {
-            if let Ok(utf8_contents) = str::from_utf8(&contents) {
-                let expected = utf8_contents.chars().count();
-                let s = String::utf8(contents);
-                s.char_len() == expected
-            } else {
-                let expected_at_most = contents.len();
-                let s = String::utf8(contents);
-                s.char_len() <= expected_at_most
-            }
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_len_binary_contents_utf8_string(contents: Vec<u8>) -> bool {
-            let expected = contents.len();
-            let s = String::utf8(contents);
-            s.len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_bytesize_binary_contents_utf8_string(contents: Vec<u8>) -> bool {
-            let expected = contents.len();
-            let s = String::utf8(contents);
-            s.bytesize() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_char_len_utf8_contents_ascii_string(contents: alloc::string::String) -> bool {
-            let expected = contents.len();
-            let s = String::ascii(contents.into_bytes());
-            s.char_len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_len_utf8_contents_ascii_string(contents: alloc::string::String) -> bool {
-            let expected = contents.len();
-            let s = String::ascii(contents.into_bytes());
-            s.len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_bytesize_utf8_contents_ascii_string(contents: alloc::string::String) -> bool {
-            let expected = contents.len();
-            let s = String::ascii(contents.into_bytes());
-            s.bytesize() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_char_len_binary_contents_ascii_string(contents: Vec<u8>) -> bool {
-            let expected = contents.len();
-            let s = String::ascii(contents);
-            s.char_len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_len_binary_contents_ascii_string(contents: Vec<u8>) -> bool {
-            let expected = contents.len();
-            let s = String::ascii(contents);
-            s.len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_bytesize_binary_contents_ascii_string(contents: Vec<u8>) -> bool {
-            let expected = contents.len();
-            let s = String::ascii(contents);
-            s.bytesize() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_char_len_utf8_contents_binary_string(contents: alloc::string::String) -> bool {
-            let expected = contents.len();
-            let s = String::binary(contents.into_bytes());
-            s.char_len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_len_utf8_contents_binary_string(contents: alloc::string::String) -> bool {
-            let expected = contents.len();
-            let s = String::binary(contents.into_bytes());
-            s.len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_bytesize_utf8_contents_binary_string(contents: alloc::string::String) -> bool {
-            let expected = contents.len();
-            let s = String::binary(contents.into_bytes());
-            s.bytesize() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_char_len_binary_contents_binary_string(contents: Vec<u8>) -> bool {
-            let expected = contents.len();
-            let s = String::binary(contents);
-            s.char_len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_len_binary_contents_binary_string(contents: Vec<u8>) -> bool {
-            let expected = contents.len();
-            let s = String::binary(contents);
-            s.len() == expected
-        }
-
-        #[allow(clippy::needless_pass_by_value)]
-        fn fuzz_bytesize_binary_contents_binary_string(contents: Vec<u8>) -> bool {
-            let expected = contents.len();
-            let s = String::binary(contents);
-            s.bytesize() == expected
-        }
-    }
+    use crate::String;
 
     #[test]
     fn center_returns_error_with_empty_padding() {
